+++ core/primitives/res/epoch_configs/mainnet/0.json
