+++ core/primitives/res/epoch_configs/mainnet/48.json
