+++ core/primitives/res/epoch_configs/mainnet/56.json
