+++ core/primitives/res/epoch_configs/mainnet/64.json
