+++ core/primitives/res/epoch_configs/mainnet/65.json
