+++ core/primitives/res/epoch_configs/mainnet/69.json
