+++ core/primitives/res/epoch_configs/mainnet/70.json
