+++ core/primitives/res/epoch_configs/mocknet/0.json
