+++ core/primitives/res/epoch_configs/mocknet/48.json
