+++ core/primitives/res/epoch_configs/mocknet/64.json
