+++ core/primitives/res/epoch_configs/mocknet/65.json
