+++ core/primitives/res/epoch_configs/mocknet/69.json
