+++ core/primitives/res/epoch_configs/mocknet/70.json
