+++ core/primitives/res/epoch_configs/testnet/0.json
