+++ core/primitives/res/epoch_configs/testnet/48.json
