+++ core/primitives/res/epoch_configs/testnet/56.json
