+++ core/primitives/res/epoch_configs/testnet/65.json
