+++ core/primitives/res/epoch_configs/testnet/69.json
