+++ core/primitives/res/epoch_configs/testnet/70.json
