+++ core/primitives/src/epoch_manager.rs
@@ -13,7 +13,9 @@ use near_primitives_core::hash::CryptoHash;
 use near_primitives_core::types::BlockHeight;
 use near_schema_checker_lib::ProtocolSchema;
 use smart_default::SmartDefault;
-use std::collections::HashMap;
+use std::collections::{BTreeMap, HashMap};
+use std::ops::Bound;
+use std::sync::Arc;
 
 pub const AGGREGATOR_KEY: &[u8] = b"AGGREGATOR";
 
@@ -89,6 +91,7 @@ pub struct AllEpochConfigTestOverrides {
 /// and returns the EpochConfig that should be used for this protocol version.
 #[derive(Clone)]
 pub struct AllEpochConfig {
+    config_store: EpochConfigStore,
     /// Whether this is for production (i.e., mainnet or testnet). This is a temporary implementation
     /// to allow us to change protocol config for mainnet and testnet without changing the genesis config
     use_production_config: bool,
@@ -108,6 +111,7 @@ impl AllEpochConfig {
         chain_id: &str,
     ) -> Self {
         Self {
+            config_store: EpochConfigStore::for_chain_id(chain_id),
             use_production_config,
             genesis_epoch_config,
             chain_id: chain_id.to_string(),
@@ -122,6 +126,7 @@ impl AllEpochConfig {
         test_overrides: Option<AllEpochConfigTestOverrides>,
     ) -> Self {
         Self {
+            config_store: EpochConfigStore::for_chain_id(chain_id),
             use_production_config,
             genesis_epoch_config,
             chain_id: chain_id.to_string(),
@@ -800,4 +805,188 @@ pub struct EpochSummary {
     /// Protocol version for next next epoch, as summary of epoch T defines
     /// epoch T+2.
     pub next_next_epoch_version: ProtocolVersion,
+}
+
+macro_rules! include_config {
+    ($chain:expr, $version:expr, $file:expr) => {
+        (
+            $chain,
+            $version,
+            include_str!(concat!(
+                env!("CARGO_MANIFEST_DIR"),
+                "/res/epoch_configs/",
+                $chain,
+                "/",
+                $file
+            )),
+        )
+    };
+}
+
+static CONFIGS: &[(&str, ProtocolVersion, &str)] = &[
+    // Epoch configs for mainnet.
+    include_config!("mainnet", 0, "0.json"),
+    include_config!("mainnet", 48, "48.json"),
+    include_config!("mainnet", 56, "56.json"),
+    include_config!("mainnet", 64, "64.json"),
+    include_config!("mainnet", 65, "65.json"),
+    include_config!("mainnet", 69, "69.json"),
+    include_config!("mainnet", 70, "70.json"),
+    // Epoch configs for testnet.
+    include_config!("testnet", 0, "0.json"),
+    include_config!("testnet", 48, "48.json"),
+    include_config!("testnet", 56, "56.json"),
+    include_config!("testnet", 64, "64.json"),
+    include_config!("testnet", 65, "65.json"),
+    include_config!("testnet", 69, "69.json"),
+    include_config!("testnet", 70, "70.json"),
+    // Epoch configs for mocknet (forknet).
+    include_config!("mocknet", 0, "0.json"),
+    include_config!("mocknet", 48, "48.json"),
+    include_config!("mocknet", 64, "64.json"),
+    include_config!("mocknet", 65, "65.json"),
+    include_config!("mocknet", 69, "69.json"),
+    include_config!("mocknet", 70, "70.json"),
+];
+
+#[derive(Clone)]
+pub struct EpochConfigStore {
+    store: BTreeMap<ProtocolVersion, Arc<EpochConfig>>,
+}
+
+impl EpochConfigStore {
+    pub fn for_chain_id(chain_id: &str) -> Self {
+        let mut store = BTreeMap::new();
+        for (chain, version, content) in CONFIGS.iter() {
+            if *chain == chain_id {
+                let config: EpochConfig = serde_json::from_str(*content).unwrap_or_else(|e| {
+                    panic!("Failed to load epoch config files for chain {}: {:#}", chain_id, e)
+                });
+                store.insert(*version, Arc::new(config));
+            }
+        }
+        Self { store }
+    }
+
+    fn get_config(&self, protocol_version: ProtocolVersion) -> &Arc<EpochConfig> {
+        self.store
+            .range((Bound::Unbounded, Bound::Included(protocol_version)))
+            .next_back()
+            .unwrap_or_else(|| {
+                panic!("Failed to find EpochConfig for protocol version {}", protocol_version)
+            })
+            .1
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::fs;
+    use std::path::Path;
+
+    use near_primitives_core::types::ProtocolVersion;
+    use near_primitives_core::version::PROTOCOL_VERSION;
+
+    use crate::epoch_manager::{AllEpochConfig, EpochConfig};
+
+    use super::EpochConfigStore;
+
+    /// Checks that stored epoch config for latest protocol version matches the
+    /// one generated by overrides from genesis config.
+    /// If the test fails, it is either a configuration bug or the stored
+    /// epoch config must be updated.
+    fn test_epoch_config_store(chain_id: &str) {
+        let genesis_epoch_config = parse_config_file(chain_id, 0).unwrap();
+        let all_epoch_config =
+            AllEpochConfig::new_with_test_overrides(true, genesis_epoch_config, chain_id, None);
+
+        let config_store = EpochConfigStore::for_chain_id(chain_id);
+        for protocol_version in 0..=PROTOCOL_VERSION {
+            let stored_config = config_store.get_config(protocol_version);
+            let expected_config = all_epoch_config.for_protocol_version(protocol_version);
+            assert_eq!(*stored_config.as_ref(), expected_config);
+        }
+    }
+
+    #[test]
+    fn test_epoch_config_store_ainnet() {
+        test_epoch_config_store("mainnet");
+    }
+
+    #[test]
+    fn test_epoch_config_store_testnet() {
+        test_epoch_config_store("testnet");
+    }
+
+    #[test]
+    fn test_epoch_config_store_mocknet() {
+        test_epoch_config_store("mocknet");
+    }
+
+    #[allow(unused)]
+    fn generate_epoch_configs(chain_id: &str) {
+        let genesis_epoch_config = parse_config_file(chain_id, 0).unwrap();
+        let all_epoch_config = AllEpochConfig::new_with_test_overrides(
+            true,
+            genesis_epoch_config.clone(),
+            chain_id,
+            None,
+        );
+
+        let mut prev_config = genesis_epoch_config;
+        for protocol_version in 1..=PROTOCOL_VERSION {
+            let next_config = all_epoch_config.for_protocol_version(protocol_version);
+            if next_config != prev_config {
+                tracing::info!("Writing config for protocol version {}", protocol_version);
+                dump_config_file(&next_config, chain_id, protocol_version);
+            }
+            prev_config = next_config;
+        }
+    }
+
+    #[test]
+    #[ignore]
+    fn generate_epoch_configs_mainnet() {
+        generate_epoch_configs("mainnet");
+    }
+
+    #[test]
+    #[ignore]
+    fn generate_epoch_configs_testnet() {
+        generate_epoch_configs("testnet");
+    }
+
+    #[test]
+    #[ignore]
+    fn generate_epoch_configs_mocknet() {
+        generate_epoch_configs("mocknet");
+    }
+
+    #[allow(unused)]
+    fn parse_config_file(chain_id: &str, protocol_version: ProtocolVersion) -> Option<EpochConfig> {
+        let path = Path::new(env!("CARGO_MANIFEST_DIR"))
+            .join("res/epoch_configs")
+            .join(chain_id)
+            .join(format!("{}.json", protocol_version));
+        if path.exists() {
+            let content = fs::read_to_string(path).unwrap();
+            let config: EpochConfig = serde_json::from_str(&content).unwrap();
+            Some(config)
+        } else {
+            None
+        }
+    }
+
+    #[allow(unused)]
+    fn dump_config_file(config: &EpochConfig, chain_id: &str, protocol_version: ProtocolVersion) {
+        let content = serde_json::to_string_pretty(config).unwrap();
+        fs::write(
+            Path::new(env!("CARGO_MANIFEST_DIR"))
+                .join("res/epoch_configs")
+                .join(chain_id)
+                .join(format!("{}.json", protocol_version)),
+            content,
+        )
+        .unwrap()
+    }
 }
